resumeai/backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pdfplumber==0.11.0
python-docx==1.1.2
spacy==3.7.4
sentence-transformers==2.7.0
scikit-learn==1.4.2
numpy==1.26.4
pydantic==1.10.13
typer==0.7.0